+++ internal/services/myplanmodifier/planmodifierdynamic/dynamic_requires_replace.go
@@ -21,17 +21,22 @@ func RequiresReplace() planmodifier.Dynamic {
 	)
 }
 
+// RequiresReplaceIfNotNull is a plan modifier that sets RequiresReplace
+// on the attribute if the planned value is different from the state value.
+// It will not trigger replacement when either the planned or state value is null.
+// When this function is called, the equality of the planned and state values
+// has already been checked by the PlanModifyDynamic method, so we can assume they are different values.
 func RequiresReplaceIfNotNull() planmodifier.Dynamic {
 	return RequiresReplaceIf(
 		func(_ context.Context, req planmodifier.DynamicRequest, resp *RequiresReplaceIfFuncResponse) {
-			if req.PlanValue.IsNull() {
+			if req.PlanValue.IsNull() || req.StateValue.IsNull() {
 				resp.RequiresReplace = false
 				return
 			}
 			resp.RequiresReplace = true
 		},
-		"If the value of this attribute changes and is not null, Terraform will destroy and recreate the resource.",
-		"If the value of this attribute changes and is not null, Terraform will destroy and recreate the resource.",
+		"If the planned value is different from the state value, Terraform will destroy and recreate the resource unless either the planned or state value is null.",
+		"If the planned value is different from the state value, Terraform will destroy and recreate the resource unless either the planned or state value is null.",
 	)
 }